+++ lib/presentation/pages/puzzle/bloc/puzzle_state.dart
@@ -23,6 +23,9 @@ abstract class PuzzleState with _$PuzzleState {
     required final Iterable<PuzzlePieceUI> pieces,
     required final PuzzlePieceUI? selectedPiece,
     required final bool isDragging,
+    required final DrawnGroup? drawnGroup,
+    required final DrawnGroupCommitStatus drawnGroupCommitStatus,
+    required final bool isDrawingGroup,
     final Offset? dragStartOffset,
     final Offset? pieceStartPosition,
     final Offset? previewPosition,
@@ -51,6 +54,9 @@ abstract class PuzzleState with _$PuzzleState {
     solutionIdx: -1,
     selectedPiece: null,
     isDragging: false,
+    drawnGroup: null,
+    drawnGroupCommitStatus: DrawnGroupCommitStatus.tooSmall,
+    isDrawingGroup: false,
     showPreview: false,
     previewCollision: false,
     moveHistory: [],
@@ -80,10 +86,7 @@ abstract class PuzzleState with _$PuzzleState {
 
   bool get isUndoEnabled => !isSolving && moveHistory.isNotEmpty && moveIndex > 0;
 
-  bool get isCustomConfig => PuzzleConfigClassifier.isCustomConfig(
-    pieces: pieces,
-    gridConfig: gridConfig,
-  );
+  bool get isCustomConfig => PuzzleConfigClassifier.isCustomConfig(pieces: pieces, gridConfig: gridConfig);
 
   bool isPieceInGrid(final String pieceId) => gridPieces.any((final e) => !e.isConfigItem && e.id == pieceId);